+++ types/module/mock_appmodule_test.go
@@ -1,3 +1,5 @@
+// module_test inconsistenctly import appmodulev2 & appmodulev1 due to limitation in mockgen
+// eventually, when we change mocking library, we should be consistent in our appmodule imports
 package module_test
 
 import (
@@ -43,5 +45,5 @@ type CoreAppModule interface {
 
 type CoreAppModuleWithPreBlock interface {
 	CoreAppModule
-	appmodule.HasPreBlocker
+	appmodulev2.HasPreBlocker
 }